Added Graphing library.

+++ core/src/main/java/com/tonny/core/blockchain/TransactionResponse.kt
@@ -10,6 +10,6 @@ data class TransactionResponse(
 )
 
 data class Coordinate(
-  val x: Double,
-  val y: Double
+  val x: Float,
+  val y: Float
 )
